refactor: api 경로 및 사이드바 컴포넌트 파일 구조 정리
