+++ apps/client/src/app/layouts/private-layout/private-layout.tsx
@@ -4,8 +4,8 @@ import { Outlet, useLocation, useNavigate, useParams } from 'react-router';
 
 import { AlertModal } from '@cds/ui';
 
-import Sidebar from '@shared/components/sidebar/sidebar';
 import { useDeleteChatRoom } from '@shared/features/memo-list-view/api/queries';
+import Sidebar from '@shared/features/sidebar/sidebar';
 import {
   LayoutUIProvider,
   useLayoutUI,
