+++ apps/client/src/pages/ai-results/apis/queries.ts
@@ -1,6 +1,6 @@
 import { useInfiniteQuery, useQuery } from '@tanstack/react-query';
 
-import { api } from '@shared/api/instance';
+import { api } from '@shared/apis/instance';
 import { type MockMemo } from '@shared/features/memo-list-view/components/memo-list/types/memo';
 import { mapApiMemoToMockMemo } from '@shared/features/memo-list-view/components/memo-list/types/memo';
 import { components } from '@shared/types/schema';
