+++ apps/client/src/pages/login/apis/exchange-google-code.ts
@@ -1,4 +1,4 @@
-import { api } from '@shared/api/instance';
+import { api } from '@shared/apis/instance';
 
 import { LOGIN_END_POINT } from './end-point';
