+++ apps/client/src/shared/features/memo-list-view/api/queries.ts
@@ -2,7 +2,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { ALL_MEMO_KEY } from '@pages/all-memo/api/query-key';
 
-import { api } from '@shared/api/instance';
+import { api } from '@shared/apis/instance';
 
 import { AI_END_POINT } from './end-point';
 import { PROMPT_END_POINT } from './end-point';
