+++ apps/client/src/shared/features/memo-list-view/components/tree-view/api/queries.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { api } from '@shared/api/instance';
+import { api } from '@shared/apis/instance';
 import { StructureMemoTypes } from '@shared/types/memo-info-type';
 
 import { STRUCTURE_MEMO_END_POINT } from './end-point';
