+++ apps/client/src/shared/features/memo-list-view/components/tree-view/components/tree-memo/api/queries.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { api } from '@shared/api/instance';
+import { api } from '@shared/apis/instance';
 import { LabelTextType } from '@shared/types/label-type';
 
 import { MEMO_MODAL_END_POINT } from './end-point';
