+++ apps/client/src/shared/features/sidebar/sidebar.css.ts
